Improved test duration
